Encapsulate UUID stuff to DRY
