+++ tests/unit/Entity/ArticleDetailTest.php
@@ -10,6 +10,7 @@ use Ramsey\Uuid\Uuid;
 /**
  * @covers \Pionyr\PionyrCz\Entity\ArticleDetail
  * @covers \Pionyr\PionyrCz\Entity\AbstractArticle
+ * @covers \Pionyr\PionyrCz\Entity\IdentifiableTrait
  * @covers \Pionyr\PionyrCz\Entity\Link
  * @covers \Pionyr\PionyrCz\Entity\Photo
  */
