+++ tests/unit/Entity/ArticlePreviewTest.php
@@ -9,6 +9,7 @@ use Ramsey\Uuid\Uuid;
 /**
  * @covers \Pionyr\PionyrCz\Entity\ArticlePreview
  * @covers \Pionyr\PionyrCz\Entity\AbstractArticle
+ * @covers \Pionyr\PionyrCz\Entity\IdentifiableTrait
  */
 class ArticlePreviewTest extends TestCase
 {
