+++ tests/unit/Entity/EventPreviewTest.php
@@ -9,6 +9,7 @@ use Ramsey\Uuid\Uuid;
 /**
  * @covers \Pionyr\PionyrCz\Entity\EventPreview
  * @covers \Pionyr\PionyrCz\Entity\AbstractEvent
+ * @covers \Pionyr\PionyrCz\Entity\IdentifiableTrait
  */
 class EventPreviewTest extends TestCase
 {
